Separate tracking and PID factories, use algorithms interface (#1417)
The original approach in #829 was to add more functionality into a
single factory. This pushes back on that by separating the two
responsibilities into two algorithms. This should improve modularity.
This also converts algorithms to the algorithms interface. The factory
code is rewritten from scratch and licensed to LGPL3.

---------
