+++ src/algorithms/pid/MatchToRICHPIDConfig.h
@@ -1,0 +1,13 @@
+// SPDX-License-Identifier: LGPL-3.0-or-later
+// Copyright (C) 2024, Dmitry Kalinkin
+
+#pragma once
+
+namespace eicrecon {
+
+struct MatchToRICHPIDConfig {
+  double etaTolerance{0.2};
+  double phiTolerance{0.1};
+};
+
+}
